requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0

# Data & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
wordcloud>=1.9.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
